backend/central/requirements.txt
Flask == 2.2.2
numpy
requests
pandas
python-dotenv
